+++ pyroomacoustics/directivities/base.py
@@ -1,43 +1,32 @@
 import abc
 
-from .direction import DirectionVector
+from .direction import DirectionVector, Rotation3D
 
 
 class Directivity(abc.ABC):
     """
     Abstract class for directivity patterns.
-
     """
 
-    def __init__(self, orientation):
-        assert isinstance(orientation, DirectionVector)
-        self._orientation = orientation
-
     @property
+    @abc.abstractmethod
     def is_impulse_response(self):
         """
-        Indicates whether the array returned has coefficients
-        for octave bands or is a full-size impulse response
+        Indicates whether the array contains coefficients for octave bands
+        (returns ``False``) or is a full-size impulse response (returns
+        ``True``).
         """
-        return False
+        raise NotImplementedError
 
-    def get_azimuth(self, degrees=True):
-        return self._orientation.get_azimuth(degrees, degrees=degrees)
-
-    def get_colatitude(self, degrees=True):
-        return self._orientation.get_colatitude(degrees, degrees=degrees)
-
-    def set_orientation(self, orientation):
+    @property
+    @abc.abstractmethod
+    def filter_len_ir(self):
         """
-        Set orientation of directivity pattern.
-
-        Parameters
-        ----------
-        orientation : DirectionVector
-            New direction for the directivity pattern.
+        When ``is_impulse_response`` returns ``True``, this property returns the
+        lengths of the impulse responses returned.
+        All impulse responses are assumed to have the same length.
         """
-        assert isinstance(orientation, DirectionVector)
-        self._orientation = orientation
+        raise NotImplementedError
 
     @abc.abstractmethod
     def get_response(
@@ -46,4 +35,4 @@ class Directivity(abc.ABC):
         """
         Get response for provided angles and frequency.
         """
-        return
+        raise NotImplementedError